+++ src/pulse3D/excel_writer.py
@@ -1,9 +1,11 @@
 # -*- coding: utf-8 -*-
 import datetime
 import logging
+import os
 from typing import Any
 from typing import List
 from typing import Tuple
+from typing import Union
 
 import numpy as np
 import pandas as pd
@@ -19,6 +21,7 @@ from .peak_detection import peak_detector
 from .plate_recording import PlateRecording
 from .plotting import plotting_parameters
 from .utils import truncate
+from .utils import truncate_float
 from .utils import xl_col_to_name
 
 
@@ -146,10 +149,7 @@ def create_frequency_vs_time_charts(
         {
             "categories": f"='{PER_TWITCH_METRICS_SHEET_NAME}'!$B${well_row + 2}:${last_column}${well_row + 2}",
             "values": f"='{PER_TWITCH_METRICS_SHEET_NAME}'!$B${well_row + 7}:${last_column}${well_row + 7}",
-            "marker": {
-                "type": "diamond",
-                "size": 7,
-            },
+            "marker": {"type": "diamond", "size": 7},
             "line": {"none": True},
         }
     )
@@ -180,49 +180,64 @@ def create_frequency_vs_time_charts(
 
 def write_xlsx(
     plate_recording: PlateRecording,
-    name: str = None,
-    start_time: float = 0.0,
-    end_time: float = np.inf,
-    twitch_widths: Tuple[int, ...] = tuple([50, 90]),
+    start_time: Union[float, int] = 0,
+    end_time: Union[float, int] = np.inf,
+    twitch_widths: Tuple[int, ...] = (50, 90),
 ):
     """Write plate recording waveform and computed metrics to Excel spredsheet.
 
     Args:
-        plate_recording (PlateRecording): loaded plate recording object
-        name (str, optional): File name of Excel spreadsheet. Defaults to None.
-        start_time (float): Start time of windowed analysis. Defaults to 0.0.
-        end_time (float): End time of windowed analysis. Defaults to np.inf.
+        plate_recording (PlateRecording): loaded PlateRecording object
+        start_time (float): Start time of windowed analysis. Defaults to 0.
+        end_time (float): End time of windowed analysis. Defaults to infinity.
 
     Raises:
         NotImplementedError: if peak finding algorithm fails for unexpected reason
-        ValueError: if start and end times are outside of expected bounds, or do not
+        ValueError: if start and end times are outside of expected bounds, or do not ?
     """
+    # make sure windows bounds are floats
+    start_time = float(start_time)
+    end_time = float(end_time)
+
     # get metadata from first well file
     w = [pw for pw in plate_recording if pw][0]
-    if name is None:
-        name = f"{w[PLATE_BARCODE_UUID]}__{w[UTC_BEGINNING_RECORDING_UUID].strftime('%Y_%m_%d_%H%M%S')}.xlsx"
     interpolated_data_period = (
         w[INTERPOLATION_VALUE_UUID] if plate_recording.is_optical_recording else INTERPOLATED_DATA_PERIOD_US
     )
 
-    # get max_time from all wells
-    max_time = max([w.force[0][-1] for w in plate_recording if w])
-    min_time = min([w.force[0][-1] for w in plate_recording if w])
-    time_points = np.arange(interpolated_data_period, max_time, interpolated_data_period)
+    # get max and min of final timepoints across each well
+    raw_timepoints = [w.force[0][-1] for w in plate_recording if w]
+    max_final_time_secs = max(raw_timepoints)
+    interpolated_timepoints_secs = np.arange(
+        interpolated_data_period, max_final_time_secs, interpolated_data_period
+    )
 
-    # Kristian 1/25/22
+    max_final_time_us = max_final_time_secs / MICRO_TO_BASE_CONVERSION
+    # produce min final time truncated to 1 decimal place
+    min_final_time_us = truncate_float(min(raw_timepoints) / MICRO_TO_BASE_CONVERSION, 1)
+
     if start_time < 0:
-        raise ValueError(f"Window start time ({start_time}) cannot be negative.")
-    if start_time >= (min_time / MICRO_TO_BASE_CONVERSION):
+        raise ValueError(f"Window start time ({start_time}s) cannot be negative")
+    if start_time >= round(min_final_time_us, 1):
         raise ValueError(
-            f"Window start time ({start_time}s) cannot be after longest time of shortest recording ({(min_time/MICRO_TO_BASE_CONVERSION):.1f})."
+            f"Window start time ({start_time}s) greater than the max timepoint of this recording ({min_final_time_us:.1f}s)"
         )
     if end_time <= start_time:
-        raise ValueError(f"Window end time ({end_time}) must come after window start time ({start_time}).")
+        raise ValueError(
+            f"Window end time ({end_time}s) must be greater than window start time ({start_time}s)"
+        )
 
-    end_time = np.min([end_time, max_time / MICRO_TO_BASE_CONVERSION])
-    is_full_analysis = np.all([(start_time == 0), (end_time == max_time / MICRO_TO_BASE_CONVERSION)])
+    end_time = min(end_time, max_final_time_us)
+    is_full_analysis = start_time == 0 and end_time == max_final_time_us
 
+    # create output file name
+    input_file_name_no_ext = os.path.splitext(os.path.basename(plate_recording.path))[0]
+    if is_full_analysis:
+        output_file_name = f"{input_file_name_no_ext}_full.xlsx"
+    else:
+        output_file_name = f"{input_file_name_no_ext}_{start_time}-{end_time}.xlsx"
+
+    # create metadata sheet format as DataFrame
     metadata = {
         "A": [
             "Recording Information:",
@@ -266,7 +281,7 @@ def write_xlsx(
             str(w[UTC_BEGINNING_RECORDING_UUID].replace(tzinfo=None)),
             "",
             "",
-            w["File Format Version"],
+            w[FILE_FORMAT_VERSION_METADATA_KEY],
             w.get(MANTARRAY_SERIAL_NUMBER_UUID, ""),
             w.get(SOFTWARE_RELEASE_VERSION_UUID, ""),
             w.get(SOFTWARE_BUILD_NUMBER_UUID, ""),
@@ -285,15 +300,14 @@ def write_xlsx(
 
     log.info("Computing data metrics for each well.")
 
-    for i, well_file in enumerate(plate_recording):
-
+    for well_file in plate_recording:
         # initialize some data structures
         error_msg = None
         peaks_and_valleys = None
         # necessary for concatenating DFs together, in event that peak-finding fails and produces empty DF
         dfs = init_dfs()
         metrics = tuple(
-            concat([dfs[k][j] for j in dfs[k].keys()], axis=1) for k in ["per_twitch", "aggregate"]
+            concat([dfs[k][j] for j in dfs[k].keys()], axis=1) for k in ("per_twitch", "aggregate")
         )
 
         if well_file is None:
@@ -304,23 +318,29 @@ def write_xlsx(
 
         # find bounding indices with respect to well recording
         well_start_idx, well_end_idx = truncate(
-            source_series=time_points, lower_bound=well_file.force[0][0], upper_bound=well_file.force[0][-1]
+            source_series=interpolated_timepoints_secs,
+            lower_bound=well_file.force[0][0],
+            upper_bound=well_file.force[0][-1],
         )
 
         # find bounding indices of specified start/end windows
         window_start_idx, window_end_idx = truncate(
-            source_series=time_points / MICRO_TO_BASE_CONVERSION, lower_bound=start_time, upper_bound=end_time
+            source_series=interpolated_timepoints_secs / MICRO_TO_BASE_CONVERSION,
+            lower_bound=start_time,
+            upper_bound=end_time,
         )
 
-        start_idx = np.max([window_start_idx, well_start_idx])
-        end_idx = np.min([window_end_idx, well_end_idx])
+        start_idx = max(window_start_idx, well_start_idx)
+        end_idx = min(window_end_idx, well_end_idx)
 
         # fit interpolation function on recorded data
         interp_data_fn = interpolate.interp1d(well_file.force[0, :], well_file.force[1, :])
 
         # interpolate, normalize, and scale data
-        interpolated_force = interp_data_fn(time_points[start_idx:end_idx])
-        interpolated_well_data = np.row_stack([time_points[start_idx:end_idx], interpolated_force])
+        interpolated_force = interp_data_fn(interpolated_timepoints_secs[start_idx:end_idx])
+        interpolated_well_data = np.row_stack(
+            [interpolated_timepoints_secs[start_idx:end_idx], interpolated_force]
+        )
 
         min_value = min(interpolated_force)
         interpolated_force -= min_value
@@ -361,23 +381,25 @@ def write_xlsx(
                 "force": interpolated_well_data,
                 "num_data_points": len(interpolated_well_data[0]),
                 "start_time": start_time,
-                "end_time": np.min([time_points[-1] / MICRO_TO_BASE_CONVERSION, end_time]),
+                "end_time": np.min([interpolated_timepoints_secs[-1] / MICRO_TO_BASE_CONVERSION, end_time]),
             }
         )
 
     # waveform table
     continuous_waveforms = {
-        "Time (seconds)": pd.Series(time_points[start_idx:end_idx] / MICRO_TO_BASE_CONVERSION)
+        "Time (seconds)": pd.Series(
+            interpolated_timepoints_secs[start_idx:end_idx] / MICRO_TO_BASE_CONVERSION
+        )
     }
     for d in data:
         continuous_waveforms[f"{d['well_name']} - Active Twitch Force (μN)"] = pd.Series(d["interp_data"])
     continuous_waveforms_df = pd.DataFrame(continuous_waveforms)
 
     _write_xlsx(
-        name=name,
-        metadata_df=metadata_df,
-        continuous_waveforms_df=continuous_waveforms_df,
-        data=data,
+        output_file_name,
+        metadata_df,
+        continuous_waveforms_df,
+        data,
         is_optical_recording=plate_recording.is_optical_recording,
         twitch_widths=twitch_widths,
     )
@@ -385,14 +407,14 @@ def write_xlsx(
 
 
 def _write_xlsx(
-    name: str,
+    output_file_name: str,
     metadata_df: pd.DataFrame,
     continuous_waveforms_df: pd.DataFrame,
     data: List[Dict[Any, Any]],
     is_optical_recording: bool = False,
-    twitch_widths: Tuple[int, ...] = tuple([50, 90]),
+    twitch_widths: Tuple[int, ...] = (50, 90),
 ):
-    with pd.ExcelWriter(name) as writer:
+    with pd.ExcelWriter(output_file_name) as writer:
         log.info("Writing H5 file metadata")
         metadata_df.to_excel(writer, sheet_name="metadata", index=False, header=False)
         ws = writer.sheets["metadata"]
@@ -413,10 +435,10 @@ def _write_xlsx(
         snapshot_sheet = wb.add_worksheet("continuous-waveform-snapshot")
         full_sheet = wb.add_worksheet("full-continuous-waveform-plots")
 
-        for i, dm in enumerate(data):
+        for well_idx, dm in enumerate(data):
             log.info(f'Creating waveform charts for well {dm["well_name"]}')
             create_waveform_charts(
-                i,
+                well_idx,
                 dm,
                 continuous_waveforms_df,
                 wb,
@@ -428,7 +450,7 @@ def _write_xlsx(
 
         # aggregate metrics sheet
         log.info("Writing aggregate metrics.")
-        aggregate_df = aggregate_metrics_df(data)
+        aggregate_df = aggregate_metrics_df(data, twitch_widths)
         aggregate_df.to_excel(writer, sheet_name="aggregate-metrics", index=False, header=False)
 
         # per twitch metrics sheet
@@ -448,7 +470,7 @@ def _write_xlsx(
 
                 num_data_points = len(dm[0])
 
-                log.info(f'Creating frequency vs time chart for well {d["well_name"]}')
+                log.info(f"Creating frequency vs time chart for well {d['well_name']}")
                 create_frequency_vs_time_charts(
                     freq_vs_time_sheet,
                     freq_vs_time_chart,
@@ -459,7 +481,7 @@ def _write_xlsx(
                     num_metrics,
                 )
 
-                log.info(f'Creating force frequency relationship chart for well {d["well_name"]}')
+                log.info(f"Creating force frequency relationship chart for well {d['well_name']}")
                 create_force_frequency_relationship_charts(
                     force_freq_sheet,
                     force_freq_chart,
@@ -468,11 +490,11 @@ def _write_xlsx(
                     num_data_points,  # number of twitches
                     num_metrics,
                 )
-        log.info(f"Writing {name}")
+        log.info(f"Writing {output_file_name}")
 
 
 def create_waveform_charts(
-    iter_idx,
+    well_idx,
     dm,
     continuous_waveforms_df,
     wb,
@@ -486,7 +508,7 @@ def create_waveform_charts(
 
     upper_x_bound = (
         dm["end_time"]
-        if (dm["end_time"] - dm["start_time"]) <= CHART_MAXIMUM_SNAPSHOT_LENGTH
+        if dm["end_time"] - dm["start_time"] <= CHART_MAXIMUM_SNAPSHOT_LENGTH
         else dm["start_time"] + CHART_MAXIMUM_SNAPSHOT_LENGTH
     )
 
@@ -560,7 +582,7 @@ def create_waveform_charts(
         waveform_charts=[snapshot_chart, full_chart],
         continuous_waveform_sheet=continuous_waveforms_sheet,
         detector_type="Peak",
-        well_index=iter_idx,
+        well_index=well_idx,
         well_name=f"{dm['well_name']}",
         upper_x_bound_cell=dm["num_data_points"],
         indices=peaks,
@@ -574,7 +596,7 @@ def create_waveform_charts(
         waveform_charts=[snapshot_chart, full_chart],
         continuous_waveform_sheet=continuous_waveforms_sheet,
         detector_type="Valley",
-        well_index=iter_idx,
+        well_index=well_idx,
         well_name=f"{dm['well_name']}",
         upper_x_bound_cell=dm["num_data_points"],
         indices=valleys,
@@ -588,42 +610,19 @@ def create_waveform_charts(
     snapshot_sheet.insert_chart(
         well_row * (CHART_HEIGHT_CELLS + 1), well_col * (CHART_FIXED_WIDTH_CELLS + 1), snapshot_chart
     )
-    full_sheet.insert_chart(1 + iter_idx * (CHART_HEIGHT_CELLS + 1), 1, full_chart)
+    full_sheet.insert_chart(1 + well_idx * (CHART_HEIGHT_CELLS + 1), 1, full_chart)
 
 
-def aggregate_metrics_df(data: List[Dict[Any, Any]], widths: Tuple[int, ...] = tuple([50, 90])):
+def aggregate_metrics_df(data: List[Dict[Any, Any]], widths: Tuple[int, ...] = (50, 90)):
     """Combine aggregate metrics for each well into single DataFrame.
 
     Args:
         data (list): list of data metrics and metadata associated with each well
-        widths (tuple of ints, optional): twitch-widths to return data for. Defaults to tuple(50,90).
+        widths (tuple of ints, optional): twitch-widths to return data for. Defaults to (50, 90).
 
     Returns:
         df (DataFrame): aggregate data frame of all metric aggregate measures
     """
-
-    def append2df(main_df: pd.DataFrame, metrics: pd.DataFrame):
-        """Append metric-specific aggregate measures to aggregate data frame.
-
-        Wraps original append function.
-
-        Args:
-            main_df (DataFrame): aggregate data frame
-            metrics (DataFrame): metric-specific aggregate measures
-        Returns:
-            main_df (DataFrame): aggregate data frame
-        """
-        metrics.reset_index(inplace=True)
-        metrics.insert(0, "level_0", [nm] + [""] * 5)
-        metrics.columns = np.arange(metrics.shape[1])
-
-        main_df = main_df.append(metrics, ignore_index=True)
-
-        # empty row
-        main_df = main_df.append(pd.Series([""]), ignore_index=True)
-
-        return main_df
-
     df = pd.DataFrame()
     df = df.append(
         pd.Series(
@@ -648,34 +647,56 @@ def aggregate_metrics_df(data: List[Dict[Any, Any]], widths: Tuple[int, ...] = t
     combined = pd.concat([d["metrics"][1] for d in data])
 
     for metric_id in ALL_METRICS:
-        if metric_id in [WIDTH_UUID, RELAXATION_TIME_UUID, CONTRACTION_TIME_UUID]:
-            for k in widths:
-                nm = CALCULATED_METRIC_DISPLAY_NAMES[metric_id].format(k)
-                metric_df = combined[metric_id][k].drop(columns=["n"]).T
-                df = append2df(df, metric_df)
+        if metric_id in (WIDTH_UUID, RELAXATION_TIME_UUID, CONTRACTION_TIME_UUID):
+            for width in widths:
+                name = CALCULATED_METRIC_DISPLAY_NAMES[metric_id].format(width)
+                metric_df = combined[metric_id][width].drop(columns=["n"]).T
+                df = _append_aggregate_measures_df(df, metric_df, name)
         else:
-            nm = CALCULATED_METRIC_DISPLAY_NAMES[metric_id]
+            name = CALCULATED_METRIC_DISPLAY_NAMES[metric_id]
             metric_df = combined[metric_id].drop(columns=["n"]).T.droplevel(level=-1, axis=0)
-            df = append2df(df, metric_df)
+            df = _append_aggregate_measures_df(df, metric_df, name)
 
     return df
 
 
-def per_twitch_df(data: List[Dict[Any, Any]], widths: Tuple[int, ...] = tuple([50, 90])):
+def _append_aggregate_measures_df(main_df: pd.DataFrame, metrics: pd.DataFrame, name: str):
+    """Append metric-specific aggregate measures to aggregate data frame.
+
+    Includes an empty row after aggregate measures
+
+    Args:
+        main_df (DataFrame): aggregate data frame
+        metrics (DataFrame): metric-specific aggregate measures
+        name (str): the display name of the metric
+    Returns:
+        main_df (DataFrame): aggregate data frame
+    """
+    metrics.reset_index(inplace=True)
+    metrics.insert(0, "level_0", [name] + [""] * 5)
+    metrics.columns = np.arange(metrics.shape[1])
+
+    main_df = main_df.append(metrics, ignore_index=True)
+
+    # empty row
+    main_df = main_df.append(pd.Series([""]), ignore_index=True)
+
+    return main_df
+
+
+def per_twitch_df(data: List[Dict[Any, Any]], widths: Tuple[int, ...] = (50, 90)):
     """Combine per-twitch metrics for each well into single DataFrame.
 
     Args:
         data (list): list of data metrics and metadata associated with each well
-        widths (tuple of ints, optional): twitch-widths to return data for. Defaults to tuple([50,90]).
+        widths (tuple of ints, optional): twitch-widths to return data for. Defaults to (50, 90).
 
     Returns:
         df (DataFrame): per-twitch data frame of all metrics
     """
-    # append to a list instead of to a dataframe directly because it's fasters
-    # construct the dataframe at the end
+    # append to a list instead of to a dataframe directly because it's faster and construct the dataframe at the end
     series_list = []
-    for j, d in enumerate(data):  # for each well
-
+    for d in data:  # for each well
         num_per_twitch_metrics = 0  # len(labels)
         twitch_times = [d["force"][0, i] / MICRO_TO_BASE_CONVERSION for i in d["metrics"][0].index]
 
@@ -687,7 +708,7 @@ def per_twitch_df(data: List[Dict[Any, Any]], widths: Tuple[int, ...] = tuple([5
         num_per_twitch_metrics += 2
 
         for metric_id in ALL_METRICS:
-            if metric_id in [WIDTH_UUID, RELAXATION_TIME_UUID, CONTRACTION_TIME_UUID]:
+            if metric_id in (WIDTH_UUID, RELAXATION_TIME_UUID, CONTRACTION_TIME_UUID):
                 for twitch_width in widths:
                     values = [f"{CALCULATED_METRIC_DISPLAY_NAMES[metric_id].format(twitch_width)}"]
                     temp = pd.Series(values + list(dm[metric_id][twitch_width]))
@@ -705,4 +726,4 @@ def per_twitch_df(data: List[Dict[Any, Any]], widths: Tuple[int, ...] = tuple([5
 
     df = pd.concat(series_list, axis=1).T
     df.fillna("", inplace=True)
-    return (df, num_per_twitch_metrics)
+    return df, num_per_twitch_metrics