+++ CHANGELOG.rst
@@ -1,5 +1,12 @@
-Changelog for Curi Bio Software Development Kit
-===============================================
+Changelog for Pulse3D
+=====================
+
+0.23.6 (2022-04-14)
+-------------------
+- Fixed issue where desired twitch widths weren't being output in aggregate metrics sheet
+- Fixed code snippets in documentation
+- Changed output file name to include input file name
+- Removed ``name`` param of ``write_xlsx`` function
 
 0.23.5 (2022-04-07)
 -------------------